move parsing functionality to own folder
