+++ src/app/utils/open-api-parser/parse/open-api-sample.ts
@@ -1,0 +1,12 @@
+import { IOpenApiResponse } from '../../../../types/open-api';
+import sample from './open-api-sample.json';
+
+export function getSample(): IOpenApiResponse {
+  return {
+    openapi: sample.openapi,
+    info: sample.info,
+    servers: sample.servers,
+    paths: sample.paths
+  };
+}
+
