--- src/app/utils/open-api-sample.ts
@@ -1,7 +1,0 @@
-import { IOpenApiResponse } from '../../types/open-api';
-import sample from './open-api-sample.json';
-
-export function getSample(): IOpenApiResponse {
-  return sample;
-}
-
